// tools/generators/hibiscus-python-library/schema.d.ts
export interface Schema {
  name: string;
  description?: string;
  tags?: string;
  directory?: string;
  customSource: boolean;
  sourceName?: string;
  sourceUrl?: string;
  sourceSecondary?: boolean;
  packageName: string;
  moduleName: string;
  publishable: boolean;
  addDevDependencies: boolean;
  buildLockedVersions: boolean;
  buildBundleLocalDependencies: boolean;
}